// api-client/models/CreateAssessmentRequestDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateAssessmentRequestDTO = {
    title?: string;
    type?: string;
    description?: string;
    maximumScore?: number;
    passingScore?: number;
    dueDate?: string;
    timeLimit?: number;
    courseId?: number;
    lessonId?: number;
};