// api-client/models/ResponseDTOAssessmentResponseDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AssessmentResponseDTO } from './AssessmentResponseDTO';
export type ResponseDTOAssessmentResponseDTO = {
    data?: AssessmentResponseDTO;
    status?: number;
    message?: string;
    errors?: Record<string, string>;
    timestamp?: string;
};